keys try 3

+++ quake/memory.h
@@ -7,8 +7,8 @@ extern "C" {
 
 	void Memory_Init();
 	char* Memory_GetData();
-	size_t Memory_GetSize();
-	size_t Memory_GetMinimum();
+	unsigned Memory_GetSize();
+	unsigned Memory_GetMinimum();
 
 #ifdef __cplusplus
 }
